@@ -1,66 +1,91 @@
 {
-	"name": "webpack-react-boilerplate",
-	"version": "1.0.0",
-	"description": "Webpack react boilerplate",
-	"main": "index.js",
-	"browserslist": [
-		"ie >= 11",
-		"last 2 version"
-	],
-	"scripts": {
-		"start": "webpack-dev-server --mode development --open --config webpack.dev.js",
-		"build": "rimraf dist && webpack --mode production  --config webpack.prod.js --progress --profile --color",
-		"analyze-bundle": "rimraf dist && webpack --mode=production --config webpack.prod.js --profile --bundleAnalyze=true --json > stats.json && webpack-bundle-analyzer stats.json dist"
-	},
-	"author": "Adam Morsi",
-	"license": "ISC",
-	"devDependencies": {
-		"@babel/core": "^7.6.4",
-		"@babel/plugin-proposal-class-properties": "^7.5.5",
-		"@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-		"@babel/plugin-syntax-dynamic-import": "^7.2.0",
-		"@babel/preset-env": "^7.6.3",
-		"@babel/preset-react": "^7.6.3",
-		"autoprefixer": "^9.7.0",
-		"babel-eslint": "^10.0.3",
-		"babel-loader": "^8.0.6",
-		"copy-webpack-plugin": "^5.0.5",
-		"css-loader": "^2.1.1",
-		"eslint": "^5.16.0",
-		"eslint-config-prettier": "^4.3.0",
-		"eslint-loader": "^2.2.1",
-		"eslint-plugin-prettier": "^3.1.1",
-		"file-loader": "^3.0.1",
-		"html-webpack-plugin": "^3.2.0",
-		"image-webpack-loader": "^6.0.0",
-		"mini-css-extract-plugin": "^0.6.0",
-		"node-sass": "^4.13.0",
-		"optimize-css-assets-webpack-plugin": "^5.0.3",
-		"postcss-loader": "^3.0.0",
-		"redux-devtools-extension": "^2.13.8",
-		"resolve-url-loader": "^3.1.0",
-		"rimraf": "^2.7.1",
-		"sass-loader": "^7.3.1",
-		"style-loader": "^0.23.1",
-		"terser-webpack-plugin": "^1.4.1",
-		"webpack": "^4.41.2",
-		"webpack-bundle-analyzer": "^3.6.0",
-		"webpack-cli": "^3.3.10",
-		"webpack-dev-server": "^3.9.0",
-		"webpack-merge": "^4.2.2"
-	},
-	"dependencies": {
-		"@hot-loader/react-dom": "^16.10.2",
-		"@tweenjs/tween.js": "^16.11.0",
-		"axios": "^0.19.0",
-		"core-js": "^3.3.5",
-		"prettier": "^1.18.2",
-		"prop-types": "^15.7.2",
-		"react": "^16.11.0",
-		"react-dom": "^16.11.0",
-		"react-hot-loader": "^4.12.15",
-		"react-router-dom": "^5.1.2",
-		"redux-logger": "^3.0.6",
-		"regenerator-runtime": "^0.13.3"
-	}
+  "name": "adam-morsi-portfolio",
+  "version": "1.0.0",
+  "description": "Adam Morsi portfolio which has information about skills, projects, etc...",
+  "engines": {
+    "node": ">=14.17.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/react-custom-projects/my-portofolio"
+  },
+  "main": "index.js",
+  "browserslist": [
+    "ie >= 11",
+    "last 2 version"
+  ],
+  "scripts": {
+    "start": "node scripts/start.js",
+    "build": "webpack --mode=production  --config buildTools/webpack.prod.js --progress --color",
+    "build:serve": "serve dist --cors -p 8080",
+    "analyze-bundle": "yarn build --json=stats.json && webpack-bundle-analyzer stats.json dist"
+  },
+  "author": "Adam Morsi",
+  "license": "ISC",
+  "devDependencies": {
+    "@babel/core": "^7.20.5",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/plugin-proposal-object-rest-spread": "^7.20.2",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.19.6",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
+    "autoprefixer": "^10.4.13",
+    "babel-loader": "^9.1.0",
+    "browserslist": "^4.21.4",
+    "clean-webpack-plugin": "^4.0.0",
+    "cli-color": "^2.0.3",
+    "copy-webpack-plugin": "^11.0.0",
+    "core-js": "^3.26.1",
+    "css-loader": "^6.7.2",
+    "css-minimizer-webpack-plugin": "^4.2.2",
+    "detect-port": "^1.5.1",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.29.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-webpack-plugin": "^3.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "jest-resolve": "^29.3.1",
+    "jest-watch-typeahead": "^2.2.1",
+    "mini-css-extract-plugin": "2.7.2",
+    "plop": "^3.1.1",
+    "postcss": "^8.4.20",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-loader": "^7.0.2",
+    "postcss-normalize": "^10.0.1",
+    "postcss-preset-env": "^7.8.3",
+    "prettier": "^2.8.1",
+    "prompts": "^2.4.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-refresh": "^0.14.0",
+    "regenerator-runtime": "^0.13.11",
+    "resolve-url-loader": "^5.0.0",
+    "rework": "^1.0.1",
+    "rework-visit": "^1.0.0",
+    "sass": "^1.56.2",
+    "sass-loader": "^13.2.0",
+    "serve": "^14.1.2",
+    "source-map-loader": "^4.0.1",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.6",
+    "webpack": "^5.75.0",
+    "webpack-bundle-analyzer": "^4.7.0",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.11.1",
+    "webpack-merge": "^5.8.0"
+  },
+  "dependencies": {
+    "react-router-dom": "^6.4.5",
+    "simplebar-react": "^2.3.0"
+  }
 }
