@@ -1,13 +1,13 @@
 import React from 'react';
 
 const Loader = () => {
-	return (
-		<div>
-			<div className="loader-wrapper">
-				<div className="loader" />
-			</div>
-		</div>
-	);
+  return (
+    <div>
+      <div className="loader-wrapper">
+        <div className="loader" />
+      </div>
+    </div>
+  );
 };
 
 export default Loader;
